+++ packages/backend/src/useCases/RefreshToken/RefreshTokenUseCase.ts
@@ -1,4 +1,5 @@
 import dayjs from 'dayjs'
+import { DevRadar_Error } from '../../errors/errors'
 
 import { GenerateTokenProvider } from '../../providers/GenerateTokenProvider'
 
@@ -13,12 +14,10 @@ export class RefreshTokenUseCase {
   ) {}
 
   async execute({ refresh_token }: IRefreshTokenDTO) {
-    const refreshToken = await this.refreshTokenRepository.findByToken(
-      refresh_token
-    )
+    const refreshToken = await this.refreshTokenRepository.findByToken({ refresh_token })
 
     if (!refreshToken) {
-      throw new Error('Refresh token invalid')
+      throw new DevRadar_Error("TOKEN_EXPIRED")
     }
 
     const refreshTokenExpired = dayjs().isAfter(
@@ -28,12 +27,10 @@ export class RefreshTokenUseCase {
     const token = this.generateTokenProvider.generator(refreshToken.userId)
 
     if (refreshTokenExpired) {
-      await this.refreshTokenRepository.deleteRefreshToken(refreshToken.userId)
+      await this.refreshTokenRepository.deleteRefreshToken({ userId: refreshToken.userId })
 
       const newRefreshToken =
-        await this.refreshTokenRepository.generateRefreshToken(
-          refreshToken.userId
-        )
+        await this.refreshTokenRepository.generateRefreshToken({ userId: refreshToken.userId })
 
       return { token, refreshToken: newRefreshToken }
     }
